requirements.txt
Django==5.1.4
djangorestframework==3.15.2
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1